+++ saleor/graphql/app/filters.py
@@ -2,10 +2,10 @@ import django_filters
 import graphene
 
 from ...app import models
-from ...app.types import AppExtensionTarget, AppType
-from ..core.descriptions import ADDED_IN_322, DEPRECATED_IN_3X_INPUT
+from ...app.types import AppType
+from ..core.descriptions import ADDED_IN_322
 from ..core.filters import EnumFilter, ListObjectTypeFilter
-from .enums import AppExtensionMountEnum, AppExtensionTargetEnum, AppTypeEnum
+from .enums import AppTypeEnum
 
 
 def filter_app_search(qs, _, value):
@@ -17,18 +17,6 @@ def filter_app_search(qs, _, value):
 def filter_app_type(qs, _, value):
     if value in [AppType.LOCAL, AppType.THIRDPARTY]:
         qs = qs.filter(type=value)
-    return qs
-
-
-def filter_app_extension_target(qs, _, value):
-    if value in [target for target, _ in AppExtensionTarget.CHOICES]:
-        qs = qs.filter(target=value)
-    return qs
-
-
-def filter_app_extension_mount(qs, _, value):
-    if value:
-        qs = qs.filter(mount__in=value)
     return qs
 
 
@@ -55,16 +43,6 @@ class AppFilter(django_filters.FilterSet):
 
 
 class AppExtensionFilter(django_filters.FilterSet):
-    mount = ListObjectTypeFilter(
-        input_class=AppExtensionMountEnum,
-        method=filter_app_extension_mount,
-        help_text=f"DEPRECATED: Use `mountName` instead. {DEPRECATED_IN_3X_INPUT}",
-    )
-    target = EnumFilter(
-        input_class=AppExtensionTargetEnum,
-        method=filter_app_extension_target,
-        help_text=f"DEPRECATED: Use `targetName` instead. {DEPRECATED_IN_3X_INPUT}",
-    )
     mountName = ListObjectTypeFilter(
         input_class=graphene.String,
         method=filter_app_extension_mount_name,
@@ -77,4 +55,4 @@ class AppExtensionFilter(django_filters.FilterSet):
 
     class Meta:
         model = models.AppExtension
-        fields = ["mount", "target", "mountName", "targetName"]
+        fields = ["mountName", "targetName"]