+++ saleor/graphql/app/tests/queries/test_app_extension.py
@@ -11,8 +11,6 @@ query ($id: ID!){
     appExtension(id: $id){
         label
         url
-        mount
-        target
         mountName
         targetName
         id
@@ -21,19 +19,6 @@ query ($id: ID!){
             code
         }
         settings
-        options {
-          ... on AppExtensionOptionsWidget{
-            widgetTarget {
-              method
-            }
-          }
-          ...on AppExtensionOptionsNewTab {
-            newTabTarget{
-              method
-            }
-          }
-
-        }
     }
 }
 """
@@ -64,15 +49,13 @@ def test_app_extension_staff_user(app, staff_api_client, permission_manage_produ
     extension_data = content["data"]["appExtension"]
     assert app_extension.label == extension_data["label"]
     assert app_extension.url == extension_data["url"]
-    assert app_extension.mount == extension_data["mount"].lower()
-    assert app_extension.target == extension_data["target"].lower()
+    assert app_extension.mount == extension_data["mountName"].lower()
+    assert app_extension.target == extension_data["targetName"].lower()
 
     assert app_extension.permissions.count() == 1
     assert len(extension_data["permissions"]) == 1
     permission_code = extension_data["permissions"][0]["code"].lower()
     assert app_extension.permissions.first().codename == permission_code
-
-    assert extension_data["options"]["widgetTarget"]["method"] == "POST"
 
     assert extension_data["settings"] is not None
     assert extension_data["settings"]["widgetTarget"]["method"] == "POST"
@@ -104,8 +87,8 @@ def test_app_extension_by_app(app, app_api_client, permission_manage_products):
     extension_data = content["data"]["appExtension"]
     assert app_extension.label == extension_data["label"]
     assert app_extension.url == extension_data["url"]
-    assert app_extension.mount == extension_data["mount"].lower()
-    assert app_extension.target == extension_data["target"].lower()
+    assert app_extension.mount == extension_data["mountName"].lower()
+    assert app_extension.target == extension_data["targetName"].lower()
 
     assert app_extension.permissions.count() == 1
     assert len(extension_data["permissions"]) == 1
@@ -313,8 +296,8 @@ query ($id: ID!){
     appExtension(id: $id){
         label
         url
-        mount
-        target
+        mountName
+        targetName
         id
         permissions{
             code
@@ -499,7 +482,7 @@ def test_app_extension_with_app_query_by_customer_without_permissions(
         (AppExtensionTarget.NEW_TAB, "GET"),
     ],
 )
-def test_app_extension_type_options(
+def test_app_extension_type_settings(
     target,
     method,
     app,
@@ -527,15 +510,11 @@ def test_app_extension_type_options(
     content = get_graphql_content(response)
     extension_data = content["data"]["appExtension"]
 
-    assert extension_data["target"] == target.upper()
-
     assert extension_data["mountName"] == "ORDER_DETAILS_WIDGETS"
     assert extension_data["targetName"] == app_extension.target.upper()
 
     if target == AppExtensionTarget.NEW_TAB:
-        assert extension_data["options"]["newTabTarget"]["method"] == method
         assert extension_data["settings"]["newTabTarget"]["method"] == method
 
     if target == AppExtensionTarget.WIDGET:
-        assert extension_data["options"]["widgetTarget"]["method"] == method
         assert extension_data["settings"]["widgetTarget"]["method"] == method